refactor: change motor interface from percentages to direct PWM values
- Changed MotorDriverInterface from setSpeed(percentage) to setPWM(0-255)
- Removed percentage conversions in AutosteerProcessor - now pure integer math
- PWM calculation: error × Kp + minPWM, clamped to highPWM
- Updated all motor drivers (PWM, Keya, Danfoss) to accept PWM values
- Removed dynamic PWM scaling that was causing incorrect values
- Fixed issue where 3° error produced 94 PWM instead of 30
- Now correctly produces expected values (e.g. 10° × Kp10 = 100 PWM)
