--- lib/aio_autosteer/MotorDriverDetector.cpp
@@ -1,5 +1,0 @@
-// MotorDriverDetector.cpp - Implementation file
-#include "MotorDriverDetector.h"
-
-// Define the static instance
-MotorDriverDetector* MotorDriverDetector::instance = nullptr;
