+++ lib/aio_system/Version.h
@@ -2,7 +2,7 @@
 #define VERSION_H
 
 // AiO New Dawn firmware version
-#define FIRMWARE_VERSION "0.1.30-alpha"
+#define FIRMWARE_VERSION "0.1.31-alpha"
 
 // Teensy board type for OTA identification
 #define TEENSY_BOARD_TYPE "Teensy41"
